Added URL redirect caching.
This will be useful if there are different URL's that redirect to the same image URL. For this case, we are currently ending up with the same image being cached multiple times.
If a redirect is requested by NSURLConnection, we store the redirect URL as a NSString in the caching system. It's key is the original URL.
The current connection is closed.
We start another download attempt in SDWebImageManager with the URL parameter set to the redirect URL. This will look into the cache to see if we have anything stored and start downloading if not.