Changed the NSOperationQueue to LIFOOperationQueue.
This provides "last in first out" functionality, which should be a big plus. The user should see the content he is currently on, and not wait for all the other images to be downloaded.
Added a slightly modified version of the LIFOOperationQueue, which is available under the MIT license and was created by Ben Harris.
Modified some aspects of SDWebImageDownloader in order to comply with the LIFOOperationQueue.
SDWebImageDownloader was changed in order to re prioritize downloads that are asked to be provided once more.