Use cache key filter when scaling an image after download
We were using the raw url as the key when trying to determine if an
image was an @2x image or not. This was a problem for me because I was
pulling different sized image from a server which did not append the @2x
to the larger images. I want to the cache key filter to append @2x to
the cache key.