Race in SDWebImageManager; occasional wrong image.
Do not execute completion block if we've been cancelled. dispatch_main_sync_safe() does not guarantee that no other code runs until our block is executed, so the operation may have been cancelled in-between. If we call the completion block after cancellation, there's a race between this load and the subsequent image load in a table view cell, for instance.