Fixes a bug on SDWebImageDownload where the callbacks can only be canceled on the first downloadImageWithURL:options:progress:completed: call for a given URL – the one that started the URL request operation.
The first call to downloadImageWithURL:options:progress:completed: for a given URL creates and return the download operation. From the second call on – the 1st operation for the same URL hasn't finished yet – no download operation is created and the callbacks are stored to be called when the URL download operation finishes.

The problem is that calls which return no download operation can't be canceled and it's callbacks will be called anyways.

The fix approach is as follows:

* SDWebImageDownloaderCancelableCallbacksOperation class is introduced
** it conforms to SDWebImageOperation protocol
** it has a cancelCallbacksBlock property
* addProgressCallback:andCompletedBlock:forURL:createCallback: now returns an instance of SDWebImageDownloaderCancelableCallbacksOperation with the cancelCallbacksBlock set to remove the callbakcs
* downloadImageWithURL:options:progress:completed: returns the download operation if it's the 1st call or the cancelableCallbacksOperation for 2nd call on

Test included:

SDWebImageDownloaderTests.m file which implements the testThatCanceledDownloadDoNotInvokeProgressBlock.