Set UIButton's placeholer-image even if the url is nil.
Yesterday，i update the SDWebImage, then i discoverred a bug which was
not existed before.I have checked the source code of SDWebImage, the
reason is that the method (- (void)sd_setImageWithURL:(nullable NSURL
*)url forState:(UIControlState)state placeholderImage:(nullable UIImage
*)placeholder) can not set the placeholder when url is nil but the old
version did. Maybe this is not a bug for you, but for us.I hope you can
consider this problem.Last, i have used this wonderful Lib for a long
time, thanks for you.