Don't unnecessarily copy image data in the Downloader Operation
In the DownloaderOperation, we were copying the image data to an NSData, then immediately unassigning the imageData property. Presumably, this was done to prevent the imageData that would be decoded from having additional data added to it while we were decoding it. However, by unassigning the only other reference to it, we already prevented that.
In the SDImageIOCoders, use NSData-initWithData to allow the class cluster to hopefully prevent an unnecessary copy of the data